fix carousel+add menu
