+++ index.php
@@ -1,6 +1,4 @@
 <?php
-    // include_once ("model/db.php");
-    // include_once ("control/auth.php");
     require('model/db.php');
 
     if (isset($_GET['controller'])) {
