+++ model/customer_db.php
@@ -24,6 +24,14 @@
         return "good";
     }
 
+    function getCustomerByPhone($phone) {
+        require('model/db.php');
+        $searchPhone = "SELECT * FROM customers WHERE phone = '$phone'";
+        $resultPhone = mysqli_query($con, $searchPhone);
+        return json_encode(mysqli_fetch_object($resultPhone));
+        
+    }
+
     function getName($phone) {
         require('model/db.php');
         $searchPhone = "SELECT * FROM customers WHERE phone = '$phone'";
