+++ view/html/UI_user/discount.php
@@ -1,0 +1,32 @@
+<div id="header-carousel" class="carousel slide" data-ride="carousel">
+    <div class="carousel-inner">
+        <div class="carousel-item active" style="height: 380px;">
+            <img class="img-fluid" src="view/images/carousel-1.jpg" alt="carousel-1">
+            <div class="carousel-caption d-none d-md-block">
+                <div class="p-3" style="font-family:Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif">
+                    <h4 class="text-uppercase text-light font-weight-large mb-3">10% off when booking</h4>
+                    <h3 class="display-2 font-weight-bold mb-4 text-light">Diverse Menu</h3>
+                    <a href="#" class="btn btn-light py-2 px-3">Order Now</a>
+                </div>
+            </div>
+        </div>
+        <div class="carousel-item" style="height: 380px;">
+            <img class="img-fluid" src="view/images/carousel-2.jpg" alt="carousel-2">
+            <div class="carousel-caption d-none d-md-block">
+                <div class="p-3 text-light" style="font-family:Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif">
+                    <h4 class="text-uppercase font-weight-large mb-3 text-light">10% off when booking</h4>
+                    <h3 class="display-2 font-weight-bold mb-4">Resonable Price</h3>
+                    <a href="#" class="btn btn-light py-2 px-3">Order Now</a>
+                </div>
+            </div>
+        </div>
+    </div>
+    <button class="carousel-control-prev" type="button" data-bs-target="#header-carousel" data-bs-slide="prev">
+        <span class="carousel-control-prev-icon" aria-hidden="true"></span>
+        <span class="visually-hidden">Previous</span>
+    </button>
+    <button class="carousel-control-next" type="button" data-bs-target="#header-carousel" data-bs-slide="next">
+        <span class="carousel-control-next-icon" aria-hidden="true"></span>
+        <span class="visually-hidden">Next</span>
+    </button>
+</div>
