+++ view/html/UI_user/home_page.php
@@ -3,28 +3,58 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Home page</title>
+    <title>Starter</title>
     <!-- ======= Styles ====== -->
       <!--  icon -->
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.2/css/all.min.css" />
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.4.1/font/bootstrap-icons.css" />
      <!--  style -->
-    <link rel="stylesheet" type="text/css" href="view/css/UI_user/style_starter.css">
     <link rel="stylesheet" type="text/css" href="view/css/UI_user/style_navbar_homepage.css">
     <link rel="stylesheet" type="text/css" href="view/bootstrap/css/bootstrap.min.css">
     <link rel="stylesheet" type="text/css" href="view/css/UI_user/home_page_reponsive.css">
-
     <!-- ======= Scripts ====== -->
-    <script src="view/bootstrap/js/bootstrap.min.js"></script>
-    <script src="view/jquery/jquery-3.6.4.js"></script>
-    <script src="view/html/UI_user/home_page.js"></script>
+    <script>
+        function getMainContent(link) {
+            $.ajax({
+                // The link we are accessing.
+                url: link,
+                    
+                // The type of request.
+                type: "get",
+                    
+                // The type of data that is getting returned.
+                dataType: "html",
+
+                success: function( strData ){
+                    document.getElementById("mainContent").innerHTML = strData;
+                    const items = document.querySelectorAll(".carousel-item");
+
+                    items.forEach((el) => {
+                        const minPerSlide = 4
+                        let next = el.nextElementSibling
+                        for (var i=1; i<minPerSlide; i++) {
+                            if (!next) {
+                                // wrap carousel by using first child
+                                next = items[0]
+                                
+                            }
+                            let cloneChild = next.cloneNode(true)
+                            el.appendChild(cloneChild.children[0])
+                            next = next.nextElementSibling
+                        }
+                    })
+                }
+            });
+        }
+    </script>
+
 </head>
-<body>
+<body onload='getMainContent("index.php?controller=user&action=getDiscount")'>
     <div class="container-fluid">
         <div class="row align-items-center py-3 pd_mobile" style="background-color: #f2f2f2;">
             <div class="col-lg-3 d-none d-lg-block px-5">
                 <a href="index.php?controller=user&action=home_page">
-                    <img src="view/images/logo.jpg" style="width: 70%;" alt="logo">
+                    <img src="/view/images/logo.jpg" style="width: 70%;" alt="logo">
                 </a>
             </div>
             <div class="col-lg-9">
@@ -67,7 +97,7 @@
                 </a>
                 <nav class="collapse show navbar" id="navbar-vertical" style="margin-top: -9px;">
                     <div class="navbar-nav w-100">
-                        <a href="/view/html/UI_user/starter.php" class="nav-item nav-link border">Khai Vị</a>
+                        <a onclick="getMainContent('index.php?controller=user&action=menu&type=starter')" class="nav-item nav-link border">Khai Vị</a>
                         <a href="#" class="nav-item nav-link border">Món Chính</a>
                         <a href="#" class="nav-item nav-link border">Tráng Miệng</a>
                         <a href="#" class="nav-item nav-link border">Đồ Ngọt</a>
@@ -78,7 +108,7 @@
             <div class="col-lg-9">
                 <nav class="navbar navbar-expand-lg bg-light navbar-light p-2" style="height: 62px;">
                     <a href="index.php?controller=user&action=home_page" class="d-block d-lg-none">
-                        <img src="view/images/logo.jpg" style="width: 100px;" alt="logo">
+                        <img src="/view/images/logo.jpg" style="width: 100px;" alt="logo">
                     </a>
                     <button type="button" 
                             ID="navbarButton"
@@ -90,12 +120,13 @@
                     </button>
                     <div class="collapse navbar-collapse justify-content-between" id="navbarCollapse" >
                         <div class="navbar-nav">
-                            <a href="index.php?controller=user&action=home_page" class="nav-item nav-link active">Trang Chủ</a>
-                            <a href="#" class="nav-item nav-link">Thực Đơn</a>
+                            <a href="index.php?controller=user&action=home_page" class="nav-item nav-link">Trang Chủ</a>
+                            <a href="#" class="nav-item nav-link active">Thực Đơn</a>
                             <a href="#" class="nav-item nav-link">Đặt Bàn</a>
                             <a href="#" class="nav-item nav-link">Tin Tức</a>
                             <a href="index.php?controller=manager&action=login" class="nav-item nav-link">Quản Trị Viên</a>
                         </div>
+                        
                         <div class="navbar-nav ml-auto nav_main">
                             <div>
                                 <a href="index.php?controller=user&action=login" class="nav-item nav-link">
@@ -110,44 +141,19 @@
                                 </a>
                             </div>
                         </div>
-                    </div>
+                    </div> 
                 </nav>
-                <div id="header-carousel" class="carousel slide" data-ride="carousel">
-                    <div class="carousel-inner">
-                        <div class="carousel-item active" style="height: 380px;">
-                            <img class="img-fluid" src="view/images/carousel-1.jpg" alt="carousel-1">
-                            <div class="carousel-caption d-none d-md-block">
-                                <div class="p-3" style="font-family:Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif">
-                                    <h4 class="text-uppercase text-light font-weight-large mb-3">10% off when booking</h4>
-                                    <h3 class="display-2 font-weight-bold mb-4 text-light">Diverse Menu</h3>
-                                    <a href="#" class="btn btn-light py-2 px-3">Order Now</a>
-                                </div>
-                            </div>
-                        </div>
-                        <div class="carousel-item" style="height: 380px;">
-                            <img class="img-fluid" src="view/images/carousel-2.jpg" alt="carousel-2">
-                            <div class="carousel-caption d-none d-md-block">
-                                <div class="p-3 text-light" style="font-family:Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif">
-                                    <h4 class="text-uppercase font-weight-large mb-3 text-light">10% off when booking</h4>
-                                    <h3 class="display-2 font-weight-bold mb-4">Resonable Price</h3>
-                                    <a href="#" class="btn btn-light py-2 px-3">Order Now</a>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                    <button class="carousel-control-prev" type="button" data-bs-target="#header-carousel" data-bs-slide="prev">
-                        <span class="carousel-control-prev-icon" aria-hidden="true"></span>
-                        <span class="visually-hidden">Previous</span>
-                    </button>
-                    <button class="carousel-control-next" type="button" data-bs-target="#header-carousel" data-bs-slide="next">
-                        <span class="carousel-control-next-icon" aria-hidden="true"></span>
-                        <span class="visually-hidden">Next</span>
-                    </button>
+                <div id="mainContent">
+                    
                 </div>
             </div>
         </div>
     </div>
+    </div>
     <!-- ======= Scripts ====== -->
-    <script src="view/script/user_navbar.js"></script>
+    <script src="/view/script/user_navbar.js"></script>
+    <!-- <script src="view/script/home_page.js"></script> -->
+    <script src="/view/bootstrap/js/bootstrap.min.js"></script>
+    <script src="/view/jquery/jquery-3.6.4.js"></script>
 </body>
 </html>
